test(app-shell): inject the write-warning sink instead of mocking sonner
The write-warning test was green alone and RED in a full run: 6 failures, all
of the shape "expected toast.warning to be called 1 times, but got 0", plus the
one assertion that checks it is NOT called passing. Deterministic, not flaky.

The file is a `.ts`, so it runs in the `unit` project — which sets
`isolate: false` (one module graph per worker, see vitest.config.mts). A
`vi.mock('sonner')` there only holds if no other file in the same worker
imported the real module first, so the provider took the real `toast` while the
test watched the mock. Every other `vi.mock('sonner')` suite in this repo is a
`.tsx` in the isolated `dom` project; this was the lone exception.

Rather than move the file to where the mock happens to work, remove the need
for one. The message builder moves to its own module, `writeWarningToast.ts`,
which imports nothing that renders and takes the toast sink as a REQUIRED
parameter — a default-to-sonner would reintroduce the very import that has to
stay out. `AdapterProvider` passes sonner's `toast`; the test passes a recorder
and asserts on it. Behaviour is unchanged, and the suite no longer depends on
worker placement or on vitest project configuration at all.

Also drops the "exported for its unit test only" escape hatch on
`toastWriteWarning` — the seam is now the parameter, so the provider keeps its
own surface, and adds a case for the no-adapter path.
